refactor: DetailsSection to use on both class and project pages
