+++ web-registry/sanity-codegen.yml
@@ -1,5 +1,5 @@
 overwrite: true
-schema: 'https://jm12rn9t.api.sanity.io/v1/graphql/staging/default'
+schema: 'https://jm12rn9t.api.sanity.io/v1/graphql/staging/details-section'
 documents:
   - 'src/graphql/sanity/*.graphql'
 generates:
