+++ web-registry/src/components/organisms/DetailsSection/DetailsSection.styles.ts
