+++ web-registry/src/components/organisms/DetailsSection/DetailsSection.types.ts
@@ -1,0 +1,10 @@
+import {
+  DetailsSection,
+  Maybe,
+  ProjectDetailsCard,
+} from 'generated/sanity-graphql';
+
+export interface DetailsSectionProps {
+  header?: Maybe<DetailsSection>;
+  credibilityCards?: Maybe<Array<Maybe<ProjectDetailsCard>>>;
+}
