+++ web-registry/src/components/organisms/DetailsSection/DetailsSection.utils.ts
